!1274 【fdol】get job summary by grpc * update the job summary watcher

+++ component/ascend-faultdiag-online/pkg/algo_src/netfault/controller/net_fault_controller.go
@@ -30,6 +30,7 @@ import (
 	"time"
 
 	"ascend-common/common-utils/hwlog"
+	"ascend-common/common-utils/utils"
 	"ascend-faultdiag-online/pkg/algo_src/netfault/algo"
 	"ascend-faultdiag-online/pkg/algo_src/netfault/controllerflags"
 	"ascend-faultdiag-online/pkg/algo_src/netfault/policy"
@@ -82,6 +83,16 @@ func markFalseDetection(superPodAbsolutePath string) {
 
 func writeNetFaultResult(result []byte, superPodPath string, curDetectionCount int) {
 	filePath := filepath.Join(superPodPath, detectionResult)
+	// check the symlink
+	isSoftlink, err := utils.IsSoftlink(filePath)
+	if err != nil {
+		hwlog.RunLog.Errorf("[NETFAULT ALGO]check file: %s is softlink or not failed: %v", filePath, err)
+		return
+	}
+	if isSoftlink {
+		hwlog.RunLog.Errorf("[NETFAULT ALGO]file: %s is symlink, unsupported", filePath)
+		return
+	}
 	/* reload will clean file */
 	if _, err := os.Stat(filePath); err == nil && !os.IsNotExist(err) && curDetectionCount == 0 {
 		if err = os.Truncate(filePath, 0); err != nil {
